cancel int16 audio

+++ source/playback.py
@@ -12,7 +12,7 @@ def get_music(midi_file):
     Load a midi file and return the PrettyMIDI object and the audio signal
     """
     music = PrettyMIDI(midi_file=midi_file)
-    waveform = music.fluidsynth().astype("int16")
+    waveform = music.fluidsynth()
     return music, waveform
